refactor(web): 拆 App.tsx —— 3209 行拆成 18 个按领域分的文件
App.tsx 是个什么都往里塞的地方：路由外壳、终端面板、会话页、文件页、登录页、
七八个弹窗、时间格式化、路径处理，还有六个别的模块反过来 `from './App'` 借的工具
函数。改一处会话列表要在三千行里找，两个人同时改必冲突。

按领域拆进 frontend/src/components/：

  terminal/  TerminalPane(929) · terminal-toolbar · SoloTerminal · tmux-keys · claude-info
  sessions/  Sessions(521) · NewSessionModal · RenameSessionModal · CloseWorktreeModal · DirPicker
  tasks/     Tasks + 派活/群发/收成三个弹窗
  files/     FilesPage      auth/ Login      session-tags · nav-icons

纯函数按内容命名单独成文件，不叫 utils：route-hash（整站唯一碰 location.hash 格式的地方）、
time-format、path-name、shell-quote。原来 shellQuote 与 shq 是同一份实现的两个拷贝，
现在只剩一个；taskNameFromPrompt 也从「夹在目录选择器中间」挪回新建会话。

App.tsx 3209 → 784 行，只剩外壳与路由。Projects/Swarm/Race/FileBrowser/WorktreePanel
不再从 './App' 借东西，改指各自的模块。

验收：typecheck / i18n / hover / vitest(253) / build 全过；浏览器实测登录、项目页、
会话页、设置页、开终端（xterm 挂上、工具条与快捷键栏都在、tmux 内容正常刷）。
