+++ frontend/src/FileBrowser.tsx
@@ -8,7 +8,7 @@ import { useI18n } from './i18n'
 import { download as p2pDownload } from './p2p/download'
 import { pathLabelKey, type P2PPathLabel } from './p2p/labels'
 import { P2PTransferStatus, type TransferView } from './p2p/P2PTransferStatus'
-import { recentDirs } from './App'
+import { recentDirs } from './components/sessions/DirPicker'
 import { usePreferences } from './preferences'
 import { dirname, fileNameOf, fmtSize, joinPath, normalizePath } from './file-utils'
 import { copyText } from './chat/blocks'
