+++ frontend/src/Projects.tsx
@@ -21,7 +21,11 @@ import { MobileSheet, SheetRow } from './shell/MobileSheet'
 import AdaptivePanel from './shell/AdaptivePanel'
 import { useLayout } from './layout'
 import { detectPrompt } from './prompt'
-import { relTime, taskNameFromPrompt, shq, NewSessionModal, DirPicker, recentDirs, pushRecentDir, CloseWorktreeModal } from './App'
+import { relTime } from './time-format'
+import { shellQuote as shq } from './shell-quote'
+import { NewSessionModal, taskNameFromPrompt } from './components/sessions/NewSessionModal'
+import { CloseWorktreeModal } from './components/sessions/CloseWorktreeModal'
+import { DirPicker, recentDirs, pushRecentDir } from './components/sessions/DirPicker'
 import { projNeeds, runningCount, waitingCount } from './project-list/project-model'
 import type { Proj, ProjSession } from './project-list/project-model'
 import { NeedsQueue, WorkbenchHead, buildNeedCards, NEEDS_CSS } from './project-list/needs-queue'
