+++ frontend/src/project-list/project-card.tsx
@@ -9,7 +9,7 @@ import { App as AntApp, Popconfirm, Tag, Tooltip } from 'antd'
 import { api } from '../api'
 import { useI18n } from '../i18n'
 import { useLayout } from '../layout'
-import { relTime } from '../App'
+import { relTime } from '../time-format'
 import { sessionLabel } from '../session-label'
 import { AgentLogo, ChevronRight, CloseIcon, FlagIcon, StarIcon, SwarmIcon } from '../icons'
 import { BranchIcon } from '../git/parts'
